+++ R/as_learner.R
@@ -34,3 +34,10 @@ as_learners.list = function(x, clone = FALSE, ...) { # nolint
 as_learners.Learner = function(x, clone = FALSE, ...) { # nolint
   list(if (clone) x$clone(deep = TRUE) else x)
 }
+
+#' @export
+#' @rdname as_learner
+as_learners.default = function(x, clone = FALSE, ...) { # nolint
+  list(as_learner(x, clone = clone, ...))
+}
+